Prevent Yocto build to stop during git patch process
`git am` command creates a commit during patch process, and this will
result in a Yocto build blocking issue if GPG singing is enabled and
passphrase is not logged into the console. To avoid this possible issue,
the `git apply` command is used instead, which does not create a commit
and prevents build to break.